refactor: 주입 방식 변경
