+++ src/main/java/racingcar/config/AppConfig.java
@@ -1,4 +1,4 @@
-package racingcar.factory;
+package racingcar.config;
 
 import racingcar.controller.RacingCarController;
 import racingcar.io.input.InputManager;
@@ -6,20 +6,22 @@ import racingcar.io.input.InputMapper;
 import racingcar.io.input.InputView;
 import racingcar.io.output.OutputView;
 import racingcar.model.MovingCondition;
+import racingcar.model.RacingCarMovingCondition;
+import racingcar.service.CarService;
 import racingcar.service.RacingCarService;
 
-public class RacingCarFactory {
+public class AppConfig {
 
     public RacingCarController racingCarController() {
         return new RacingCarController(inputManager(), outputView(), racingCarService());
     }
 
-    private RacingCarService racingCarService() {
+    private CarService racingCarService() {
         return new RacingCarService(movingCondition());
     }
 
     private MovingCondition movingCondition() {
-        return new MovingCondition();
+        return new RacingCarMovingCondition();
     }
 
     private OutputView outputView() {
